feat: add pinned forms to suggested forms spot
